+++ docs/changes.md
@@ -5,6 +5,7 @@
   * 修复enums编译报错问题(感谢智明提供补丁)
   * 删除demouiold的1m资源让stm32的板子可以编译过(感谢智明提供补丁)
   * 修复启用WITHOUT_FSCRIPT后无法给widget设置自定义属性的问题(感谢智明提供补丁)
+  * 增加tk_strnlen和wcsncmp函数让stm32项目编译通过(感谢智明提供补丁)
 
 2022/10/08
   * 修复systembar上有edit时窗口复位的问题(issue #773)。